+++ src/aux-common/bots/test/BotActionsTests.ts
@@ -28,6 +28,8 @@ import {
     webhook,
     reject,
     html,
+    loadFile,
+    saveFile,
 } from '../BotEvents';
 import {
     COMBINE_ACTION_NAME,
@@ -3314,6 +3316,74 @@ export function botActionsTests(
                             def: 123,
                         },
                     }),
+                ]);
+            });
+        });
+
+        describe('server.loadFile()', () => {
+            it('should issue a LoadFileAction in a remote event', () => {
+                const state: BotsState = {
+                    thisBot: {
+                        id: 'thisBot',
+                        tags: {
+                            abc: true,
+                            'test()': 'server.loadFile("path")',
+                        },
+                    },
+                };
+
+                // specify the UUID to use next
+                uuidMock.mockReturnValue('uuid-0');
+                const botAction = action('test', ['thisBot']);
+                const result = calculateActionEvents(
+                    state,
+                    botAction,
+                    createSandbox
+                );
+
+                expect(result.hasUserDefinedEvents).toBe(true);
+
+                expect(result.events).toEqual([
+                    remote(
+                        loadFile({
+                            path: 'path',
+                        })
+                    ),
+                ]);
+            });
+        });
+
+        describe('server.saveFile()', () => {
+            it('should issue a SaveFileAction in a remote event', () => {
+                const state: BotsState = {
+                    thisBot: {
+                        id: 'thisBot',
+                        tags: {
+                            abc: true,
+                            'test()':
+                                'server.saveFile("path", mod.export({ abc: true }))',
+                        },
+                    },
+                };
+
+                // specify the UUID to use next
+                uuidMock.mockReturnValue('uuid-0');
+                const botAction = action('test', ['thisBot']);
+                const result = calculateActionEvents(
+                    state,
+                    botAction,
+                    createSandbox
+                );
+
+                expect(result.hasUserDefinedEvents).toBe(true);
+
+                expect(result.events).toEqual([
+                    remote(
+                        saveFile({
+                            path: 'path',
+                            data: JSON.stringify({ abc: true }),
+                        })
+                    ),
                 ]);
             });
         });